requirements.txt
rich>=13.0.0
openai>=1.0.0
click>=8.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pyperclip>=1.8.0
watchdog>=3.0.0
pytest>=7.0.0
typer>=0.9.0
pathlib2>=2.3.0 